unify open vault transaction receipts
