+++ features/borrow/open/pipes/openVaultTransactions.ts
@@ -9,9 +9,9 @@ import { jwtAuthGetToken } from 'features/termsOfService/jwt'
 import { transactionToX } from 'helpers/form'
 import { zero } from 'helpers/zero'
 import { Observable, of } from 'rxjs'
-import Web3 from 'web3'
 
 import { TxError } from '../../../../helpers/types'
+import { parseVaultIdFromReceiptLogs } from '../../../shared/transactions'
 import { OpenVaultChange, OpenVaultState } from './openVault'
 
 type ProxyChange =
@@ -171,24 +171,6 @@ export function applyOpenVaultTransaction(
   return state
 }
 
-interface Receipt {
-  logs: { topics: string[] | undefined }[]
-}
-
-export function parseVaultIdFromReceiptLogs({ logs }: Receipt): BigNumber | undefined {
-  const newCdpEventTopic = Web3.utils.keccak256('NewCdp(address,address,uint256)')
-  return logs
-    .filter((log) => {
-      if (log.topics) {
-        return log.topics[0] === newCdpEventTopic
-      }
-      return false
-    })
-    .map(({ topics }) => {
-      return new BigNumber(Web3.utils.hexToNumber(topics![3]))
-    })[0]
-}
-
 export function openVault(
   { sendWithGasEstimation }: TxHelpers,
   change: (ch: OpenVaultChange) => void,
@@ -220,7 +202,7 @@ export function openVault(
           )
 
           // assume that user went through ToS flow and can interact with application
-          const jwtToken = jwtAuthGetToken(account as string)
+          const jwtToken = jwtAuthGetToken(account)
           if (id && jwtToken) {
             saveVaultUsingApi$(
               id,
