+++ features/borrow/open/tests/openVault.test.ts
@@ -13,7 +13,7 @@ import { zero } from 'helpers/zero'
 import { of, Subject } from 'rxjs'
 import { map } from 'rxjs/operators'
 
-import { parseVaultIdFromReceiptLogs } from '../pipes/openVaultTransactions'
+import { parseVaultIdFromReceiptLogs } from '../../../shared/transactions'
 import { newCDPTxReceipt } from './fixtures/newCDPtxReceipt'
 
 describe('openVault', () => {
@@ -52,7 +52,7 @@ describe('openVault', () => {
           ilk: 'ETH-Z',
         }),
       )
-      expect(state).to.throw()
+      expect(state).to.throw(`Ilk ETH-Z does not exist`)
     })
 
     it('should start by default at the editing stage', () => {
