+++ features/earn/guni/manage/pipes/guniActionsCalls.ts
@@ -12,10 +12,10 @@ import { TxHelpers } from '../../../../../components/AppContext'
 import { transactionToX } from '../../../../../helpers/form'
 import { zero } from '../../../../../helpers/zero'
 import { GuniToken } from '../../../../../types/ethers-contracts'
-import { parseVaultIdFromReceiptLogs } from '../../../../borrow/open/pipes/openVaultTransactions'
 import { Quote } from '../../../../exchange/exchange'
 import { VaultType } from '../../../../generalManageVault/vaultType'
 import { ManageChange } from '../../../../multiply/manage/pipes/manageMultiplyVaultTransactions'
+import { parseVaultIdFromReceiptLogs } from '../../../../shared/transactions'
 import { saveVaultUsingApi$ } from '../../../../shared/vaultApi'
 import { jwtAuthGetToken } from '../../../../termsOfService/jwt'
