+++ features/shared/transactions.ts
@@ -1,0 +1,20 @@
+import { BigNumber } from 'bignumber.js'
+import Web3 from 'web3'
+
+interface Receipt {
+  logs: { topics: string[] | undefined }[]
+}
+
+export function parseVaultIdFromReceiptLogs({ logs }: Receipt): BigNumber | undefined {
+  const newCdpEventTopic = Web3.utils.keccak256('NewCdp(address,address,uint256)')
+  return logs
+    .filter((log) => {
+      if (log.topics) {
+        return log.topics[0] === newCdpEventTopic
+      }
+      return false
+    })
+    .map(({ topics }) => {
+      return new BigNumber(Web3.utils.hexToNumber(topics![3]))
+    })[0]
+}
